pyproject.toml
[project]
name = "dataops-unifor"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Alves",email = "lipealvessax@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymongo (>=4.12.1,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
